Fix frame switching and menu persistence
Fix special key for toggling menu persistence, and make sure
one-key-update-buffer-contents is called after window size is changed
so that it displays properly if the one-key buffer is designated as
a special buffer to open in a new frame.
Also set action for <frame-switch> event to select previous frame in
one-key-command.

Altered items:
one-key-special-keybindings, one-key-command, one-key-set-window-state